refactor(habitat): move workspace dtos into service model
